+++ Client/src/Remote/Endpoints/CategoryEndpoint2.tsx
@@ -1,12 +1,10 @@
-import React from 'react'
 import { get } from "../RemoteHelper"
 
-
 interface ICategoryModel {
-    id: number,
-    name: string
+  id: number,
+  name: string
 }
 export const listCategories = async (): Promise<ICategoryModel[]> => {
-    const categories = await get('/category')
-    return categories
+  const categories = await get('/category')
+  return categories
 }
