+++ Client/src/Remote/Endpoints/MaterialEndpoint2.tsx
@@ -1,12 +1,13 @@
 import { get } from "../RemoteHelper"
+
 interface IMaterialModel {
-    details: string,
-    composition: string,
-    id: number
+  details: string,
+  composition: string,
+  id: number
 }
 
 
 export const listMaterials = async (): Promise<IMaterialModel[]> => {
-    const materials = await get('/material')
-    return materials
+  const materials = await get('/material')
+  return materials
 }
