+++ Client/src/Remote/Endpoints/SubcategoryEndpoint2.tsx
@@ -1,12 +1,11 @@
-import React from 'react'
 import { get } from "../RemoteHelper"
 
 interface ISubcategoryModel {
-    id: number,
-    name: string
+  id: number,
+  name: string
 }
 
 export const listSubcategories = async (): Promise<ISubcategoryModel[]> => {
-    const subcategories = await get('/subcategory')
-    return subcategories
+  const subcategories = await get('/subcategory')
+  return subcategories
 }
